Prepping test scenarios for a richer heat map.
